refactor: refactoring Canvas Page
