+++ apps/frontend/src/features/drawing/components/index.ts
@@ -1,1 +1,4 @@
-export * from './ERDDiagramTool';
+export * from './TableNode';
+export * from './RelationshipMarker';
+export * from './Toolbar';
+export * from './EdgeSelector';
