+++ apps/frontend/src/features/drawing/hooks/index.ts
@@ -1,1 +1,3 @@
 export * from './useDragAndDrop';
+export * from './useEdges';
+export * from './useNodes';
