+++ apps/frontend/src/features/drawing/index.ts
@@ -1,1 +1,3 @@
 export * from './components';
+export * from './hooks';
+export * from './types';
