+++ apps/frontend/src/pages/index.ts
@@ -1,1 +1,2 @@
 export { LandingPage } from './LandingPage';
+export { CanvasPage } from './CanvasPage';
